core/web/entity: remove webhook and project-integration code superseded by Actions
