+++ backend/core/src/ci/reporter.rs
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: AGPL-3.0-only
  */
 
-use crate::ci::webhook::validate_webhook_url;
+use crate::ci::http_validation::validate_webhook_url;
 use anyhow::{Context, Result};
 use async_trait::async_trait;
 use serde::{Deserialize, Serialize};
