+++ backend/core/src/types/entity_aliases.rs
@@ -60,14 +60,11 @@ pub type EProject = project::Entity;
 pub type EProjectAction = project_action::Entity;
 pub type EProjectActionDelivery = project_action_delivery::Entity;
 pub type EProjectFlakeInputOverride = project_flake_input_override::Entity;
-pub type EProjectIntegration = project_integration::Entity;
 pub type EProjectTrigger = project_trigger::Entity;
 pub type ERole = role::Entity;
 pub type ESession = session::Entity;
 pub type EUploadSession = upload_session::Entity;
 pub type EUser = user::Entity;
-pub type EWebhook = webhook::Entity;
-pub type EWebhookDelivery = webhook_delivery::Entity;
 pub type EWorkerRegistration = worker_registration::Entity;
 
 pub type MApi = api::Model;
@@ -100,14 +97,11 @@ pub type MProject = project::Model;
 pub type MProjectAction = project_action::Model;
 pub type MProjectActionDelivery = project_action_delivery::Model;
 pub type MProjectFlakeInputOverride = project_flake_input_override::Model;
-pub type MProjectIntegration = project_integration::Model;
 pub type MProjectTrigger = project_trigger::Model;
 pub type MRole = role::Model;
 pub type MSession = session::Model;
 pub type MUploadSession = upload_session::Model;
 pub type MUser = user::Model;
-pub type MWebhook = webhook::Model;
-pub type MWebhookDelivery = webhook_delivery::Model;
 pub type MWorkerRegistration = worker_registration::Model;
 
 pub type AApi = api::ActiveModel;
@@ -140,14 +134,11 @@ pub type AProject = project::ActiveModel;
 pub type AProjectAction = project_action::ActiveModel;
 pub type AProjectActionDelivery = project_action_delivery::ActiveModel;
 pub type AProjectFlakeInputOverride = project_flake_input_override::ActiveModel;
-pub type AProjectIntegration = project_integration::ActiveModel;
 pub type AProjectTrigger = project_trigger::ActiveModel;
 pub type ARole = role::ActiveModel;
 pub type ASession = session::ActiveModel;
 pub type AUploadSession = upload_session::ActiveModel;
 pub type AUser = user::ActiveModel;
-pub type AWebhook = webhook::ActiveModel;
-pub type AWebhookDelivery = webhook_delivery::ActiveModel;
 pub type AWorkerRegistration = worker_registration::ActiveModel;
 
 pub type CApi = api::Column;
@@ -180,14 +171,11 @@ pub type CProject = project::Column;
 pub type CProjectAction = project_action::Column;
 pub type CProjectActionDelivery = project_action_delivery::Column;
 pub type CProjectFlakeInputOverride = project_flake_input_override::Column;
-pub type CProjectIntegration = project_integration::Column;
 pub type CProjectTrigger = project_trigger::Column;
 pub type CRole = role::Column;
 pub type CSession = session::Column;
 pub type CUploadSession = upload_session::Column;
 pub type CUser = user::Column;
-pub type CWebhook = webhook::Column;
-pub type CWebhookDelivery = webhook_delivery::Column;
 pub type CWorkerRegistration = worker_registration::Column;
 
 // `R*` (Relation) aliases removed - sea-orm relations are referenced via the
