+++ backend/entity/src/lib.rs
@@ -36,13 +36,10 @@ pub mod project;
 pub mod project_action;
 pub mod project_action_delivery;
 pub mod project_flake_input_override;
-pub mod project_integration;
 pub mod project_trigger;
 pub mod role;
 pub mod server;
 pub mod session;
 pub mod upload_session;
 pub mod user;
-pub mod webhook;
-pub mod webhook_delivery;
 pub mod worker_registration;
