+++ backend/test-support/src/fakes/mod.rs
@@ -12,5 +12,4 @@ pub mod job_reporter;
 pub mod mock_server;
 pub mod nix_store;
 pub mod store_fixture;
-pub mod webhooks;
 pub mod worker_store;
