+++ backend/test-support/src/lib.rs
@@ -35,9 +35,8 @@ pub mod prelude {
     pub use crate::fakes::job_reporter::{RecordingJobReporter, ReportedEvent};
     pub use crate::fakes::mock_server::{MockProtoServer, MockServerConn};
     pub use crate::fakes::store_fixture::{StoreFixture, load_store};
-    pub use crate::fakes::webhooks::RecordingWebhookClient;
     pub use crate::fakes::worker_store::FakeWorkerStore;
     pub use crate::fixtures::*;
     pub use crate::log_storage::{NoopLogStorage, RecordingLogStorage};
-    pub use crate::state::{test_state, test_state_recorded, test_state_with_log_storage};
+    pub use crate::state::{test_state, test_state_with_log_storage};
 }
