+++ backend/web/tests/cache_local_priority.rs
@@ -9,7 +9,6 @@
 
 use axum::extract::connect_info::MockConnectInfo;
 use axum_test::TestServer;
-use gradient_core::ci::WebhookClient;
 use gradient_core::storage::{EmailSender, NarStore};
 use gradient_core::types::ids::*;
 use gradient_core::types::{ServerState, WebDb, WorkerDb};
@@ -17,7 +16,6 @@ use sea_orm::{DatabaseBackend, MockDatabase};
 use std::net::SocketAddr;
 use std::sync::Arc;
 use test_support::fakes::email::InMemoryEmailSender;
-use test_support::fakes::webhooks::RecordingWebhookClient;
 use test_support::log_storage::NoopLogStorage;
 use test_support::prelude::test_cli;
 use uuid::Uuid;
@@ -69,9 +67,7 @@ fn build_server(cache: entity::cache::Model, peer: &str) -> TestServer {
         config: Arc::new(
             gradient_core::types::RuntimeConfig::from_cli(&cli).expect("valid test config"),
         ),
-        log_storage: Arc::new(NoopLogStorage),
-        webhooks: Arc::new(RecordingWebhookClient::new()) as Arc<dyn WebhookClient>,
-        email: Arc::new(InMemoryEmailSender::new()) as Arc<dyn EmailSender>,
+        log_storage: Arc::new(NoopLogStorage),        email: Arc::new(InMemoryEmailSender::new()) as Arc<dyn EmailSender>,
         nar_storage,
         manifest_state: Arc::new(std::sync::Mutex::new(std::collections::HashMap::new())),
         pending_credentials: Arc::new(std::sync::Mutex::new(std::collections::HashMap::new())),
