+++ backend/web/tests/evals_artefacts.rs
@@ -14,7 +14,6 @@ use axum::http::StatusCode;
 use axum_test::TestServer;
 use entity::build::BuildStatus;
 use entity::evaluation::EvaluationStatus;
-use gradient_core::ci::WebhookClient;
 use gradient_core::storage::{EmailSender, NarStore};
 use gradient_core::types::ids::*;
 use gradient_core::types::{ServerState, WebDb, WorkerDb};
@@ -22,7 +21,6 @@ use sea_orm::{DatabaseBackend, MockDatabase};
 use serde_json::Value;
 use std::sync::Arc;
 use test_support::fakes::email::InMemoryEmailSender;
-use test_support::fakes::webhooks::RecordingWebhookClient;
 use test_support::log_storage::NoopLogStorage;
 use test_support::prelude::test_cli;
 use uuid::Uuid;
@@ -224,9 +222,7 @@ fn make_state(db: sea_orm::DatabaseConnection) -> Arc<ServerState> {
         config: Arc::new(
             gradient_core::types::RuntimeConfig::from_cli(&cli).expect("valid test config"),
         ),
-        log_storage: Arc::new(NoopLogStorage),
-        webhooks: Arc::new(RecordingWebhookClient::new()) as Arc<dyn WebhookClient>,
-        email: Arc::new(InMemoryEmailSender::new()) as Arc<dyn EmailSender>,
+        log_storage: Arc::new(NoopLogStorage),        email: Arc::new(InMemoryEmailSender::new()) as Arc<dyn EmailSender>,
         nar_storage,
         manifest_state: Arc::new(std::sync::Mutex::new(std::collections::HashMap::new())),
         pending_credentials: Arc::new(std::sync::Mutex::new(std::collections::HashMap::new())),
