+++ backend/web/tests/forge_hooks.rs
@@ -17,7 +17,7 @@
 
 use axum_test::TestServer;
 use entity::evaluation::EvaluationStatus;
-use gradient_core::ci::{WebhookClient, encrypt_webhook_secret};
+use gradient_core::ci::actions::encrypt_secret_with_file as encrypt_webhook_secret;
 use gradient_core::storage::{EmailSender, NarStore};
 use gradient_core::types::ids::*;
 use gradient_core::types::triggers::TriggerConfig;
@@ -29,7 +29,6 @@ use sha2::Sha256;
 use std::sync::Arc;
 use test_support::cli::test_cli_with_crypt;
 use test_support::fakes::email::InMemoryEmailSender;
-use test_support::fakes::webhooks::RecordingWebhookClient;
 use test_support::log_storage::NoopLogStorage;
 use test_support::prelude::test_cli;
 use uuid::Uuid;
@@ -73,9 +72,7 @@ fn make_state(
         web_db: WebDb::new(db),
         worker_db: WorkerDb::new(MockDatabase::new(DatabaseBackend::Postgres).into_connection()),
         config: std::sync::Arc::new(gradient_core::types::RuntimeConfig::from_cli(&cli).expect("valid test config")),
-        log_storage: Arc::new(NoopLogStorage),
-        webhooks: Arc::new(RecordingWebhookClient::new()) as Arc<dyn WebhookClient>,
-        email: Arc::new(InMemoryEmailSender::new()) as Arc<dyn EmailSender>,
+        log_storage: Arc::new(NoopLogStorage),        email: Arc::new(InMemoryEmailSender::new()) as Arc<dyn EmailSender>,
         nar_storage,
         manifest_state: Arc::new(std::sync::Mutex::new(std::collections::HashMap::new())),
         pending_credentials: Arc::new(std::sync::Mutex::new(std::collections::HashMap::new())),
