+++ backend/web/tests/metrics.rs
@@ -12,7 +12,6 @@
 //! result set per expected scrape.
 
 use axum_test::TestServer;
-use gradient_core::ci::WebhookClient;
 use gradient_core::storage::{EmailSender, NarStore};
 use gradient_core::types::{
     MetricsConfig, RuntimeConfig, SecretString, ServerState, WebDb, WorkerDb,
@@ -22,7 +21,6 @@ use std::collections::BTreeMap;
 use std::sync::Arc;
 use test_support::cli::test_cli;
 use test_support::fakes::email::InMemoryEmailSender;
-use test_support::fakes::webhooks::RecordingWebhookClient;
 use test_support::log_storage::NoopLogStorage;
 use web::create_router;
 
@@ -55,9 +53,7 @@ fn state_with_metrics(enabled: bool, db: DatabaseConnection) -> Arc<ServerState>
         web_db: WebDb::new(db),
         worker_db: WorkerDb::new(MockDatabase::new(DatabaseBackend::Postgres).into_connection()),
         config: Arc::new(runtime),
-        log_storage: Arc::new(NoopLogStorage),
-        webhooks: Arc::new(RecordingWebhookClient::new()) as Arc<dyn WebhookClient>,
-        email: Arc::new(InMemoryEmailSender::new()) as Arc<dyn EmailSender>,
+        log_storage: Arc::new(NoopLogStorage),        email: Arc::new(InMemoryEmailSender::new()) as Arc<dyn EmailSender>,
         nar_storage,
         manifest_state: Arc::new(std::sync::Mutex::new(std::collections::HashMap::new())),
         pending_credentials: Arc::new(std::sync::Mutex::new(std::collections::HashMap::new())),
