+++ backend/web/tests/projects_sign_cache.rs
@@ -23,7 +23,6 @@ use serde_json::Value;
 use std::sync::Arc;
 use test_support::cli::test_cli;
 use test_support::fakes::email::InMemoryEmailSender;
-use test_support::fakes::webhooks::RecordingWebhookClient;
 use test_support::fixtures::{org, org_id, project_id, test_date, user, user_id};
 use test_support::log_storage::NoopLogStorage;
 use uuid::Uuid;
@@ -78,10 +77,7 @@ fn make_server(db: sea_orm::DatabaseConnection) -> TestServer {
         web_db: WebDb::new(db),
         worker_db: WorkerDb::new(MockDatabase::new(DatabaseBackend::Postgres).into_connection()),
         config,
-        log_storage: Arc::new(NoopLogStorage),
-        webhooks: Arc::new(RecordingWebhookClient::new())
-            as Arc<dyn gradient_core::ci::WebhookClient>,
-        email: Arc::new(InMemoryEmailSender::new()) as Arc<dyn EmailSender>,
+        log_storage: Arc::new(NoopLogStorage),        email: Arc::new(InMemoryEmailSender::new()) as Arc<dyn EmailSender>,
         nar_storage,
         manifest_state: Arc::new(std::sync::Mutex::new(std::collections::HashMap::new())),
         pending_credentials: Arc::new(std::sync::Mutex::new(std::collections::HashMap::new())),
