+++ backend/web/tests/roles_crud.rs
@@ -444,79 +444,3 @@ fn delete_unused_custom_role_succeeds() {
     });
 }
 
-// ── Custom-role authorization ────────────────────────────────────────────────
-
-/// A custom role with `ManageWebhooks` set must satisfy a webhook-management
-/// authorization gate, exercising the bitmask path through `access::has_permission`.
-#[test]
-fn custom_role_with_manage_webhooks_can_list_webhooks() {
-    run(async {
-        let session_id = SessionId::now_v7();
-        let token = make_token(session_id);
-        let custom_id = RoleId::now_v7();
-        let custom_membership = organization_user::Model {
-            id: OrganizationUserId::now_v7(),
-            organization: org_id(),
-            user: user_id(),
-            role: custom_id,
-        };
-        let custom_role = role::Model {
-            id: custom_id,
-            name: "webhook-ops".into(),
-            organization: Some(org_id()),
-            permission: Permission::ViewOrg.bit() | Permission::ManageWebhooks.bit(),
-            managed: false,
-        };
-
-        let db = with_auth(MockDatabase::new(DatabaseBackend::Postgres), session_id)
-            .append_query_results([vec![org()]])
-            .append_query_results([vec![custom_membership]])
-            .append_query_results([vec![custom_role]])
-            .append_query_results::<entity::webhook::Model, _, _>([
-                Vec::<entity::webhook::Model>::new(),
-            ]);
-
-        let server = make_test_server(db.into_connection());
-        let res = server
-            .get("/api/v1/webhook/test-org")
-            .add_header("authorization", format!("Bearer {}", token))
-            .await;
-
-        res.assert_status_ok();
-    });
-}
-
-#[test]
-fn custom_role_without_manage_webhooks_is_rejected() {
-    run(async {
-        let session_id = SessionId::now_v7();
-        let token = make_token(session_id);
-        let custom_id = RoleId::now_v7();
-        let custom_membership = organization_user::Model {
-            id: OrganizationUserId::now_v7(),
-            organization: org_id(),
-            user: user_id(),
-            role: custom_id,
-        };
-        let custom_role = role::Model {
-            id: custom_id,
-            name: "view-only".into(),
-            organization: Some(org_id()),
-            permission: Permission::ViewOrg.bit(),
-            managed: false,
-        };
-
-        let db = with_auth(MockDatabase::new(DatabaseBackend::Postgres), session_id)
-            .append_query_results([vec![org()]])
-            .append_query_results([vec![custom_membership]])
-            .append_query_results([vec![custom_role]]);
-
-        let server = make_test_server(db.into_connection());
-        let res = server
-            .get("/api/v1/webhook/test-org")
-            .add_header("authorization", format!("Bearer {}", token))
-            .await;
-
-        res.assert_status(axum::http::StatusCode::FORBIDDEN);
-    });
-}
